autotools: rely on `DEBUGBUILD` to enable checksrc
Instead of `CURLDEBUG`, which did work in most cases because debug
builds (almost?) always enable `CURLDEBUG`, but it's not accurate.

Unless this was chosen intentionally for some reason I'm missing.
(E.g. to enable this for non-debug builds with `--enable-curldebug`?
But this option means `Enable curl debug memory tracking`, which
seems to suggest otherwise.)