cmake: allow disabling `ENABLE_CURLDEBUG` with `ENABLE_DEBUG=ON`
Before this patch, `ENABLE_CURLDEBUG` was unconditionally enabled
when `ENABLE_DEBUGBUILD` was set. This made testing various build
configurations complicated. To fix this, after this patch
`ENABLE_CURLDEBUG` will only be set if not already defined by the
user.

This allows to use this config:
ENABLE_DEBUGBUILD=ON
ENABLE_CURLDEBUG=OFF

to enable debug features, without also enabling memory tracking.

This is important because some other build methods allow to set
one of these features but not the other. This patch allows to test
any combination with CMake.

This makes it unnecessary to use the workaround of passing
`-DDEBUGBUILD` via `CMAKE_C_FLAGS`. This also had the disadvabtage
that the CMake logic couldn't easily detect it, e.g. for disabling
symbol hiding on Windows.